+++ nest/core/decorators/__init__.py
@@ -1,2 +1,3 @@
-from nest.core.decorators.controller import Controller, Get, Post, Delete, Put, Patch
+from nest.core.decorators.controller import Controller
+from nest.core.decorators.http_methods import Get, Post, Delete, Put, Patch
 from nest.core.decorators.database import db_request_handler, async_db_request_handler
